% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econum-package.R
\docType{package}
\name{econum-package}
\alias{econum-package}
\title{Various Tools Developed at EcoNum Lab}
\description{
Control of chemostats and titrators. Calculation routines for EcoNum lab.
}
\section{Important functions}{

\itemize{
\item \code{\link[=new_econum_data]{new_econum_data()}} to manage EcoNum data,
\item \code{\link[=wtw_open]{wtw_open()}} to control WTW devices.
\item \code{\link[=iks_open]{iks_open()}} to control an IKS Aquastar,
\item \code{\link[=gce8_open]{gce8_open()}} to control a GCE 8-relay card,
}
}

