% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iks.R
\name{iks}
\alias{iks}
\alias{print.iksdata}
\alias{iks_process}
\alias{iks_get_all}
\alias{iks_get_name}
\alias{iks_get_data}
\alias{iks_set_data}
\alias{iks_open}
\alias{iks_close}
\alias{iks_cmd}
\alias{iks_info}
\alias{iks_get_config}
\title{Control the IKS Aquastar computer from within R}
\usage{
\method{print}{iksdata}(x, ...)

iks_process(msg, iks_name = "iks")

iks_get_all()

iks_get_name(port = getOption("iks_port"))

iks_get_data(port = getOption("iks_port"), iks_name = NULL)

iks_set_data(data, port = getOption("iks_port"), iks_name = NULL)

iks_open(
  port = getOption("iks_port"),
  iks_name = "iks",
  probes_names = paste("p0", 1:8, sep = ""),
  sockets_names = sprintf("s\%02d", 1:16),
  procfun = iks_process
)

iks_close(port = getOption("iks_port"))

iks_cmd(code, port = getOption("iks_port"), new_line = TRUE)

iks_info(port = getOption("iks_port"), wait = 500)

iks_get_config(port = getOption("iks_port"), wait = 500)
}
\arguments{
\item{x}{An \code{iksdata} object.}

\item{...}{Further arguments passed to the \code{print()} method.}

\item{msg}{The message send back from the IKS Aquastar.}

\item{iks_name}{The name you want to use for the IKS Aquastar module.}

\item{port}{A serial port such as \code{"com1"}, or its number such as \code{1}. For
\code{iks_close()}, you can also specify \code{"all"} to close all IKS ports.}

\item{data}{The IKS data to record (this is a list with components name,
port, version, config, probes, units, triggers, sockets, sockets_ctrl and time.}

\item{probes_names}{The name to use for the eight IKS probes.}

\item{sockets_names}{The name to use for the 16 IKS sockets.}

\item{procfun}{An \R function used to process data returned by the IKS
Aquastar.}

\item{code}{Code to send to the IKS Aquastar.}

\item{new_line}{Do we add a newline after the code send (yes by default)?}

\item{wait}{Time to wait between characters send to the IKS Aquastar (500ms
by default). Do not use too low values, or the device will not receive all
code correctly.}
}
\value{
\code{iks_open()} returns \code{TRUE} in case of success, \code{FALSE} otherwise.

\code{iks_close()} returns the com port(s) that was/were closed.

\code{iks_cmd()} and \code{iks_info()} return nothing. They are used for their side
effects of triggering some response by the IKS Aquastar. That response is
treated in an asynchronous way by the \code{procfun} provided to \code{iks_open()}.

\code{iks_get_all()} gets the list of all currently connected IKS Aquastars (a
vector of characters with corresponding com ports, and with names as
specified in the \code{iks_name} argument of \code{iks_open()}).

\code{iks_get_name()} gets the name of an IKS Aquastar, providing the com port it
is connected to.

\code{iks_get_data()} and \code{iks_set_data()} are used to retrieve or set the current
state of the IKS. \code{iks_set_data()} should not be used in a different context
than an \code{iks_process()} and should update the IKS data according to what is
received from the IKS Aquastar trough the com port.

\code{iks_get_config()} triggers a dumping of the complete configuration of the
IKS Aquastar... it takes a couple of seconds to complete! After this
command, the whole current config is stored in binary format in the \code{config}
component and the current state of sockets is updated.

\code{iks_process()} is the default workhorse function to process incoming IKS
data. It can be replaced by you own custom function, but it is already pretty
flexible. The default function allows to call custom events in the form of R
functions that accept two arguments, \code{data} and \code{iks_name}. They must be
saved in the \code{TempEnv} environment and have predefined names, depending on
the event that should trigger them: \code{iks_date_proc()} when the date is
received from the IKS, indicating a complete cycle of data sending,
\code{iks_socket_proc()} on receiving socket config data, and \code{iks_config_proc()}
on receiving IKS config data.
}
\description{
Connect to an IKS Aquastar through a serial (com) port and control it.
}
\examples{
\dontrun{
iks_get_all()
iks_open(port = "com1") # Change to the port where your IKS is connected!
iks_get_all()
iks_get_name()

# To debug exchanges with the IKS aquastar:
options(debug_iks = TRUE)

iks_get_data()   # Current data as recorded in the IKS
iks_get_config() # Get the whole IKS config
iks_get_data()   # Should be visible now!

# Stop the connection with the IKS
iks_close()
iks_get_all()
}
}
\seealso{
\code{\link[=wtw_open]{wtw_open()}}, \code{\link[=gce8_open]{gce8_open()}}
}
\author{
Philippe Grosjean \email{Philippe.Grosjean@umons.ac.be}
}
\concept{hardware control}
\keyword{utilities}
