% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{options_econum}
\alias{options_econum}
\alias{get_opt_econum}
\alias{set_opt_econum}
\title{Manage EcoNum options}
\usage{
options_econum(opt)

get_opt_econum(key, default = NULL)

set_opt_econum(key, value)
}
\arguments{
\item{opt}{A list with all EcoNum options to set. Optional and if not
provided, a default list is created at first use of \code{options_econum()}.}

\item{key}{A character string with an option name.}

\item{default}{Default (optional) value to return for the option, in case it
is not defined.}

\item{value}{Anything to set as an option in the list.}
}
\value{
For \code{options_econum()}, the whole list of options is returned. For
the other functions, associated value (or default value) is returned.
}
\description{
Examine, set or retrieve EcoNum-specific options.
}
\examples{
# Remote and local EcoNumData repositories
get_opt_econum("local_repos")
get_opt_econum("remote_repos")

# Default general metadata
get_opt_econum("def_project")
get_opt_econum("def_sample")
get_opt_econum("def_sample_date")
get_opt_econum("def_author")

# Get some non-existing EcoNum option
get_opt_econum("nokey", default = "my default value")
# Define it
set_opt_econum("nokey", "some data")
get_opt_econum("nokey", default = "my default value")
# Eliminate this key
set_opt_econum("nokey", NULL)
get_opt_econum("nokey", default = "my default value")
}
\seealso{
\link{EcoNumData}, \code{\link[=time_to_fingerprint]{time_to_fingerprint()}}
}
\author{
Philippe Grosjean \email{Philippe.Grosjean@umons.ac.be}
}
\keyword{utilities}
