% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{time_to_fingerprint}
\alias{time_to_fingerprint}
\alias{fingerprint_to_time}
\title{Calculate fingerprint from time object, or the opposite}
\usage{
time_to_fingerprint(Time)

fingerprint_to_time(hexmode, tz = "")
}
\arguments{
\item{Time}{A POSIXct object indicating time an \code{EcoNumData} object is
created.}

\item{hexmode}{An hexmode object, or a character string that can be converted
to an hexmode object.}

\item{tz}{The time zone in which to recreate the time object, by default, use
current time zone for this machine.}
}
\value{
A character string with hexadecimal representation of the time. This
could be used as a unique identifier (fingerprint) for objects that are
sequencially created in time. The reverse function gets time from a
fingerprint.
}
\description{
Get an hexadecimal unique identifier corresponding to time an object is
created. A time object from a hexadecimal fingerprint can be also obtained.
}
\examples{
(fp <- time_to_fingerprint(Sys.time()))
fingerprint_to_time(fp) # Reverse process
rm(fp)
}
\seealso{
\code{\link[=new_econum_data]{new_econum_data()}}, \code{\link[=options_econum]{options_econum()}}
}
\author{
\email{Philippe.Grosjean@umons.ac.be}
}
\keyword{utilities}
