% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtw.R
\name{wtw}
\alias{wtw}
\alias{wtw_open}
\alias{wtw_close}
\alias{wtw_ready}
\alias{wtw_sim}
\alias{wtw_cmd}
\alias{wtw_echo}
\alias{wtw_info}
\alias{wtw_read}
\alias{wtw_record}
\title{Control a WTW instrument of series P3, P4, 340, 340i or 197i}
\usage{
wtw_open(port = getOption("wtw_port"))

wtw_close(port = getOption("wtw_port"))

wtw_ready(port = getOption("wtw_port"))

wtw_sim()

wtw_cmd(code, port = getOption("wtw_port"))

wtw_echo(port = getOption("wtw_port"))

wtw_info(port = getOption("wtw_port"), timeout = 1000)

wtw_read(port = getOption("wtw_port"), timeout = 1000, O2 = FALSE)

wtw_record(file = NULL, interval = 10, Nmax = 10,
  port = getOption("wtw_port"), timeout = 1000, O2 = TRUE,
  init.data = NULL, graph = TRUE, title = "WTW record", ...)
}
\arguments{
\item{port}{A serial port as \code{"com1"}, or the port number as \code{1}.}

\item{code}{Code to send to the WTW instrument, i.e., \code{"K.1"} -> \code{"K.19"}
to simulate keystrokes (simple or double), and \code{"D.0"} -> \code{"D.12"} to read
display items (see the WTW communication protocole manual).}

\item{timeout}{Maximum time to wait in ms for the answer from the instrument.}

\item{O2}{For instruments that support it, do we also read O2 values?}

\item{file}{The file where to write the records. If \code{NULL}, a default
name is computed which is \code{'WTW_<date>.txt'} in the temporary directory. If
it is \code{NA}, nothing is written in a file.}

\item{interval}{Interval between two measurements, in second.}

\item{Nmax}{Maximum number of readings to record.}

\item{init.data}{A data frame with data to start with (append data to it).}

\item{graph}{Do we update a graph each time new item is read?}

\item{title}{The title of the graph.}

\item{...}{Further graph arguments.}
}
\value{
\code{wtw_open()} returns \code{TRUE} in case of success, \code{FALSE} otherwise.
\code{wtw_ready()} does the same if the instrument is ready or not.

\code{wtw_close()} returns the com port that was closed.

\code{wtw_info} reads the identifier of the WTW instrument and returns it (if it
is one of the recognized devices).

\code{wtw_echo()} echo data send by the instrument to the console (mainly for
debugging purpose). Use \code{wtw_read()} instead to read measurements.

\code{wtw_sim()} simulates data read from a fake instrument. You can define your
own function, if you like.

\code{wtw_record()} records a series of measurement on a disk file, and possibly
plot the data in real time. Note that this is the only function that does not
require that you open (\code{wtw_open()}) before, and close (\code{wtw_close()}) the
port before interacting with the instrument.
}
\description{
Connect to a compatible WTW instrument using a serial (com) port through the
control cable AK340/B and send commands and/or read values of pH, O2, etc.
}
\note{
A fake instrument for test and set up of control programs is available
if you specify \code{"com0"} as serial port. Note that, for simulations, you can
speed up globally data acquisition by defining \code{option(econum_speed = X)}. If
\code{'X'} is higher that 1, it means everything is done faster (\code{X = 10} means
ten times faster). If \code{'X'} is lower that 1, everything is acquired at a
slower pace: \code{X = 0.1} means ten times slower.

Do not forget to close the port after use!
}
\examples{
# Use the fake WTW instrument on com0
oport <- options(wtw_port = "com0") # Change this to connect to a real instrument
wtw_open()
wtw_echo()	# Echo data send by run/enter from the instrument
wtw_read()  # Read data from the instrument
wtw_ready() # Check that instrument is ready
wtw_close()
wtw_ready()	# Not ready any more, because we closed the port

# Record 10 data with an interval of 1 sec and make a graph...
# but speed up things by ten times if we use a fake instrument
if (getOption("wtw_port") == "com0") options(econum_speed = 10)
# Record ten data points (and do not save to a file)
# Note that you don't have to open/close the port with this function!
wtw_record(file = NA, interval = 1, Nmax = 10)
# Restore default speed
options(econum_speed = NULL)
# Restore port
options(oport); rm(oport)
rm(wtw_dat)
}
\seealso{
\code{\link[=iks_open]{iks_open()}}, \code{\link[=gce8_open]{gce8_open()}}
}
\author{
Philippe Grosjean \email{Philippe.Grosjean@umons.ac.be}
}
\concept{
Scientific instruments control
}
\keyword{utilities}
